fix: cast enum to uint8_t in ternary expressions for GCC compatibility
GCC with -Wextra/-Werror rejects ternary expressions mixing uint8_t (desc)
and compact_state::descriptor enum values. Clang accepts this silently.

Fix: static_cast<uint8_t> on the enum branch at three sites in game_state.cpp:
- make_built_group_move (line 584)
- init_payload_and_hash (line 999)
- update_card_descriptor caller (line 1275)

+++ src/main/game/search-state/game_state.cpp
@@ -582,7 +582,7 @@ void game_state::make_built_group_move(move m) {
             bottom_cid, parent_cid, rules.build_pol,
             foundations_base, rules.max_rank);
         new_desc = (desc != 0) ? desc
-            : compact_state::ROOT;
+            : static_cast<uint8_t>(compact_state::ROOT);
     }
     update_card_descriptor(bottom_cid, new_desc);
 
@@ -997,7 +997,7 @@ void game_state::init_payload_and_hash() {
                     cid, parent_cid, rules.build_pol,
                     foundations_base, rules.max_rank);
                 new_desc = (desc != 0) ? desc
-                    : compact_state::ROOT;
+                    : static_cast<uint8_t>(compact_state::ROOT);
             }
             update_card_descriptor(cid, new_desc);
         }
@@ -1272,7 +1272,7 @@ compact_state game_state::recompute_payload_from_scratch() const {
                     uint8_t desc = parent_table::get_descriptor_for_parent(
                         cid, parent_cid, rules.build_pol,
                         foundations_base, rules.max_rank);
-                    new_desc = (desc != 0) ? desc : compact_state::ROOT;
+                    new_desc = (desc != 0) ? desc : static_cast<uint8_t>(compact_state::ROOT);
                 }
             }
             cp.set_descriptor(cid, new_desc);
